api/requirements.txt
Flask
mysql-connector-python
Flask-Cors